test: cover runner cutoff scheduling
